requirements.txt
Flask==2.2.5
scikit-learn==1.3.2
pandas==2.1.2
numpy==1.26.0
joblib==1.3.2
python-dotenv==1.0.0
openai==1.12.0
pytesseract==0.3.10
Pillow==10.0.0
opencv-python-headless==4.8.1.78